requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.0.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
